Add sample picker
